Move src/js/rbac and src/js/auth into auth directory
