+++ src/chrome/create-chrome.ts
@@ -1,8 +1,8 @@
 import qe from '../utils/iqeEnablement';
 import { bootstrap, chromeInit } from './entry';
 import initializeJWT from '../jwt/initialize-jwt';
-import { createFetchPermissionsWatcher } from '../js/rbac/fetchPermissions';
-import { LibJWT } from '../js/auth';
+import { createFetchPermissionsWatcher } from '../auth/fetchPermissions';
+import { LibJWT } from '../auth';
 import { ChromeAPI, ChromeUser } from '@redhat-cloud-services/types';
 import { Store } from 'redux';
